Don't set max value for bar charts
The `data_max_len` that was being used for the maximum value of bars in
a bar chart was being used for both the number of bars to show on the
chart and the value represented by a full bar. This meant that for the
for any charts where the values were higher than the number of bars, all
bars appeared the same in the chart.

The maximum value for a bar chart is handled by the BarChart component
in tui so setting it is not required.