Remove some more clones / allocations (#30)
* style(table): use "row_highlight_style" instead of "highlight_style"

to avoid deprecation warning

* deps(tuirealm): update to 2.1.0

* refactor(components::span::Span::view): use "as_text_span" instead of manual impl

re 913b0807586e1c426f0363870e633a2bee3c3958
Also changed behavior to not panic, and instead skip any values that are not "TextSpan"s

* refactor(components::span::Span::view): avoid further clones when getting text

* refactor(components::paragraph::Paragraph::view): avoid clones when getting text

* refactor(components::label::Label::view): avoid clones when getting text

* refactor(components::list::List::view): avoid clones when getting text

* style(components::phantom): fix reference to "Spinner"

* refactor(components::radio::Radio::view): avoid clones when getting choices

* refactor(components::select::Select::view): avoid clones when getting choices

* refactor(components::table::Table::view): avoid clones when getting rows

* refactor(components::table::Table::view): avoid clones when getting headers

also remove "Table::headers" as it is only ever written to and never actually read. (compared to a scoped-variable)

* refactor(components::table::Table::view): avoid clones when getting highlight symbol

also remove "Table::hg_str" as it is only ever written to and never actually read. (compared to a scoped-variable)

* refactor(components::textarea::Textarea::view): avoid clones when getting highlight symbol

also remove "Textarea::hg_str" as it is only ever written to and never actually read. (compared to a scoped-variable)

* refactor(components::textarea::Textarea::view): avoid clones when getting text